Move TestCall to test module
