+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/helpers/QueryChannelsPostponeHelperTests.kt
@@ -7,9 +7,9 @@ import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import io.getstream.chat.android.client.Mother
 import io.getstream.chat.android.client.api.ChatApi
-import io.getstream.chat.android.client.call.asCall
 import io.getstream.chat.android.client.clientstate.ClientState
 import io.getstream.chat.android.client.clientstate.ClientStateService
+import io.getstream.chat.test.asCall
 import org.amshove.kluent.When
 import org.amshove.kluent.calling
 import org.amshove.kluent.shouldBeEqualTo
