+++ stream-chat-android-test/src/main/java/io/getstream/chat/test/CallTestUtils.kt
@@ -1,0 +1,8 @@
+package io.getstream.chat.test
+
+import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.utils.Result
+
+public fun <T : Any> callFrom(valueProvider: () -> T): Call<T> = TestCall(Result(valueProvider()))
+
+public fun <T : Any> T.asCall(): Call<T> = TestCall(Result(this))
